Refactor layout files
