+++ src/components/Header/Header.module.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/_variables";
-@import "../../../styles/breakpoints.module";
+@import "../../styles/variables";
+@import "../../styles/breakpoints.module";
 
 .Header {
   position: relative;
