+++ src/components/Header/UserWidget/UserWidget.module.scss
@@ -1,4 +1,4 @@
-@import "../../../../styles/variables";
+@import "../../../styles/variables";
 
 
 .UserWidget {
