+++ src/components/Header/UserWidget/UserWidget.tsx
