+++ src/layouts/MainLayout/MainLayout.tsx
@@ -1,7 +1,7 @@
 import useMediaQuery from "@mui/material/useMediaQuery";
 import style from "./MainLayout.module.scss";
 import {useTheme} from "@mui/material";
-import {Header} from "@/components/Layout/Header/Header";
+import {Header} from "@/components/Header/Header";
 
 export const MainLayout = ({children}: any) => {
     const theme = useTheme();
@@ -16,10 +16,10 @@ export const MainLayout = ({children}: any) => {
             />
             <div className={style.Main}>
                 <div className={style.MainWrapper}>
+                  <div className={style.MonitorContent}>
+                    {children}
+                  </div>
                     <div className={style.MonitorFrame}></div>
-                    <div className={style.MonitorContent}>
-                        {children}
-                    </div>
                 </div>
             </div>
         </div>
