+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import type {Page} from "@/types/page";
 import {Login} from "@/components/Login/Login";
-import {MainLayout} from "@/components/Layout/MainLayout";
+import {MainLayout} from "@/components/../layouts/MainLayout";
 import {useAuthConfig} from "@/services/auth";
 import {Game} from "@/components/Game";
